requirements.txt
Django==1.11.13
djangorestframework==3.9.0
psycopg2==2.7.5
pytz==2018.5
